Solve patch problem
When providing a patch to a updateFoo mutation, it is unclear whether a field whose value is null should be updated to null or left unchanged.
This is because the patch is a partial object, and, by design of GraphQL, the null value is not distinguishable from the absence of a value.